@@ -1,6 +1,7 @@
 {
   "name": "{{ name }}",
   "version": "1.0.0",
+  "type": "module",
   "description": "{{{ description }}}",
   "author": "{{{ author }}}",
   "license": "ISC",
@@ -16,35 +17,33 @@
     "build": "tsc",
     "dev": "nodemon --exec \"npm start\"",
     "start": "probot run ./lib/index.js",
-    "lint": "standard **/*.ts --fix",
-    "test": "jest && standard **/*.ts",
+    "lint": "eslint \"{src,test}/**/*.{ts,tsx,js}\"",
+    "lint:fix": "npm run lint -- --fix",
+    "test": "jest",
     "test:watch": "jest --watch --notify --notifyMode=change --coverage"
   },
   "dependencies": {
     "probot": "^7.0.0"
   },
   "devDependencies": {
-    "@types/jest": "^23.1.5",
+    "@eslint/js": "^9.37.0",
+    "@types/jest": "^30.0.0",
     "@types/node": "^10.5.2",
+    "@typescript-eslint/eslint-plugin": "^8.46.0",
+    "@typescript-eslint/parser": "^8.46.0",
+    "eslint": "^9.37.0",
+    "eslint-config-prettier": "^10.1.8",
     "eslint-plugin-typescript": "^0.12.0",
-    "jest": "^23.4.0",
+    "globals": "^16.4.0",
+    "jest": "^30.2.0",
     "nodemon": "^1.17.2",
+    "prettier": "^3.6.2",
     "smee-client": "^1.0.2",
-    "standard": "^10.0.3",
-    "ts-jest": "^23.0.0",
-    "typescript": "^2.9.2",
-    "typescript-eslint-parser": "^18.0.0"
+    "ts-jest": "^29.4.4",
+    "typescript": "^5.9.3",
+    "typescript-eslint": "^8.46.0"
   },
   "engines": {
     "node": ">= 8.3.0"
-  },
-  "standard": {
-    "parser": "typescript-eslint-parser",
-    "env": [
-      "jest"
-    ],
-    "plugins": [
-      "typescript"
-    ]
   }
 }
